Remove import cycles
