+++ src/crypto/mode/cfb.js
@@ -24,8 +24,7 @@
 
 import { AES_CFB } from '@openpgp/asmcrypto.js/dist_es8/aes/cfb';
 import * as stream from '@openpgp/web-stream-tools';
-import { getCipher } from '../crypto';
-import * as cipher from '../cipher';
+import getCipher from '../cipher/getCipher';
 import util from '../../util';
 import enums from '../../enums';
 
@@ -62,7 +61,8 @@ export async function encrypt(algo, key, plaintext, iv, config) {
     return aesEncrypt(algo, key, plaintext, iv, config);
   }
 
-  const cipherfn = new cipher[algoName](key);
+  const Cipher = getCipher(algo);
+  const cipherfn = new Cipher(key);
   const block_size = cipherfn.blockSize;
 
   const blockc = iv.slice();
@@ -104,7 +104,8 @@ export async function decrypt(algo, key, ciphertext, iv) {
     return aesDecrypt(algo, key, ciphertext, iv);
   }
 
-  const cipherfn = new cipher[algoName](key);
+  const Cipher = getCipher(algo);
+  const cipherfn = new Cipher(key);
   const block_size = cipherfn.blockSize;
 
   let blockp = iv;
