+++ src/crypto/public_key/elliptic/ecdh.js
@@ -31,7 +31,7 @@ import util from '../../../util';
 import { b64ToUint8Array } from '../../../encoding/base64';
 import * as pkcs5 from '../../pkcs5';
 import { keyFromPublic, keyFromPrivate, getIndutnyCurve } from './indutnyKey';
-import { getCipher } from '../../crypto';
+import getCipher from '../../cipher/getCipher';
 
 const webCrypto = util.getWebCrypto();
 const nodeCrypto = util.getNodeCrypto();
